+++ fixtures/coverall/tests/bindings/test_coverall.kts
@@ -214,6 +214,11 @@ Coveralls("test_regressions").use { coveralls ->
     assert(coveralls.getStatus("success") == "status: success")
 }
 
+Coveralls("test_empty_records").use { coveralls ->
+    assert(coveralls.setAndGetEmptyStruct(EmptyStruct()) == EmptyStruct())
+    assert(EmptyStruct() !== EmptyStruct())
+}
+
 class KotlinGetters : Getters {
     override fun getBool(v: Boolean, arg2: Boolean) : Boolean {
         return v != arg2
@@ -419,4 +424,49 @@ assert(d.integer == 42UL)
 // Test bytes
 Coveralls("test_bytes").use { coveralls ->
     assert(coveralls.reverse("123".toByteArray(Charsets.UTF_8)).toString(Charsets.UTF_8) == "321")
+}
+
+// Test fakes using open classes
+
+class FakePatch(private val color: Color): Patch(NoPointer) {
+    override fun `getColor`(): Color = color
+}
+
+class FakeCoveralls(private val name: String) : Coveralls(NoPointer) {
+    private val repairs = mutableListOf<Repair>()
+
+    override fun `addPatch`(patch: Patch) {
+        repairs += Repair(Instant.now(), patch)
+    }
+
+    override fun `getRepairs`(): List<Repair> {
+        return repairs
+    }
+}
+
+FakeCoveralls("using_fakes").use { coveralls ->
+    val patch = FakePatch(Color.RED)
+    coveralls.addPatch(patch)
+    assert(!coveralls.getRepairs().isEmpty())
+}
+
+FakeCoveralls("using_fakes_and_calling_methods_without_override_crashes").use { coveralls ->
+    var exception: Throwable? = null
+    try {
+        coveralls.cloneMe()
+    } catch (e: Throwable) {
+        exception = e
+    }
+    assert(exception != null)
+}
+
+Coveralls("using_fakes_with_real_objects_crashes").use { coveralls ->
+    val patch = FakePatch(Color.RED)
+    var exception: Throwable? = null
+    try {
+        coveralls.addPatch(patch)
+    } catch (e: Throwable) {
+        exception = e
+    }
+    assert(exception != null)
 }