+++ uniffi_bindgen/src/bindings/ruby/test.rs
@@ -34,7 +34,14 @@ pub fn test_script_command(
     let test_helper = UniFFITestHelper::new(fixture_name)?;
     let out_dir = test_helper.create_out_dir(tmp_dir, &script_path)?;
     let cdylib_path = test_helper.copy_cdylib_to_out_dir(&out_dir)?;
-    generate_bindings(&cdylib_path, None, &[TargetLanguage::Ruby], &out_dir, false)?;
+    generate_bindings(
+        &cdylib_path,
+        None,
+        &[TargetLanguage::Ruby],
+        None,
+        &out_dir,
+        false,
+    )?;
 
     let rubypath = env::var_os("RUBYLIB").unwrap_or_else(|| OsString::from(""));
     let rubypath = env::join_paths(